refactor: modularize ingest pipeline
