+++ schema_manager.py
@@ -1,0 +1,353 @@
+from typing import Dict
+from pyspark.sql.types import (BooleanType, DateType, DecimalType, DoubleType, StringType, StructField, StructType, TimestampType)
+# -----------------------------------------------------------------------------
+class SchemaManager:
+    """Stores all schema definitions and column mappings."""
+
+    TABLE_SCHEMAS: Dict[str, StructType] = {
+        # ... (the full schema definitions for lead, lead_xref, lead_assignment) ...
+        "lead": StructType([
+            StructField("STG_LCR_LEAD_KEY", StringType(), True),
+            StructField("LEAD_GUID", StringType(), True),
+            StructField("LEGACY_LEAD_ID", StringType(), True),
+            StructField("INDIV_ID", StringType(), True),
+            StructField("HH_ID", StringType(), True),
+            StructField("ADDR_ID", StringType(), True),
+            StructField("LEAD_CODE", StringType(), True),
+            StructField("LEAD_TYPE_ID", DecimalType(38, 0), True),
+            StructField("LEAD_TYPE", StringType(), True),
+            StructField("LEAD_SOURCE", StringType(), True),
+            StructField("LEAD_CREATE_DATE", TimestampType(), True),
+            StructField("FIRST_NAME", StringType(), True),
+            StructField("MIDDLE_NAME", StringType(), True),
+            StructField("LAST_NAME", StringType(), True),
+            StructField("SUFFIX", StringType(), True),
+            StructField("BIRTH_DATE", StringType(), True),
+            StructField("AGE", DecimalType(38, 0), True),
+            StructField("SEX", StringType(), True),
+            StructField("STREET_1", StringType(), True),
+            StructField("STREET_2", StringType(), True),
+            StructField("CITY", StringType(), True),
+            StructField("STATE_ID", DecimalType(38, 0), True),
+            StructField("STATE", StringType(), True),
+            StructField("ZIP", StringType(), True),
+            StructField("ZIP5", StringType(), True),
+            StructField("COUNTY", StringType(), True),
+            StructField("COUNTRY", StringType(), True),
+            StructField("PHONE", StringType(), True),
+            StructField("HOME_PHONE", StringType(), True),
+            StructField("CELL_PHONE", StringType(), True),
+            StructField("WORK_PHONE", StringType(), True),
+            StructField("DO_NOT_CALL", StringType(), True),
+            StructField("CALLER_ID", StringType(), True),
+            StructField("EMAIL", StringType(), True),
+            StructField("DYNAMIC_LEAD", StringType(), True),
+            StructField("PROSPECT_ID", StringType(), True),
+            StructField("EXT_PARTNER_ID", StringType(), True),
+            StructField("CHANNEL_ID", DecimalType(38, 0), True),
+            StructField("CHANNEL", StringType(), True),
+            StructField("OPT_SOURCE_ID", StringType(), True),
+            StructField("SOURCE_ID", DecimalType(38, 0), True),
+            StructField("SUB_SOURCE_ID", BooleanType(), True),
+            StructField("SOURCE_OF_REFERRAL", StringType(), True),
+            StructField("DIVISION", StringType(), True),
+            StructField("LEAD_SUB_SOURCE", StringType(), True),
+            StructField("LEAD_SUB_SOURCE_ID", StringType(), True),
+            StructField("LENDER", StringType(), True),
+            StructField("LOAN_AMOUNT", StringType(), True),
+            StructField("LOAN_DATE", DateType(), True),
+            StructField("DIABETES", StringType(), True),
+            StructField("HEALTH_PROBLEMS", StringType(), True),
+            StructField("HEART_PROBLEMS", StringType(), True),
+            StructField("HEIGHT", StringType(), True),
+            StructField("HIGH_BP_CHOL", StringType(), True),
+            StructField("IS_INSURED", StringType(), True),
+            StructField("SMOKER", StringType(), True),
+            StructField("OCCUPATION", StringType(), True),
+            StructField("SPOUSE", StringType(), True),
+            StructField("COBORROWER_AGE", DoubleType(), True),
+            StructField("COBORROWER_BIRTH_DATE", TimestampType(), True),
+            StructField("COBORROWER_HEIGHT", StringType(), True),
+            StructField("COBORROWER_ON_MORTGAGE", StringType(), True),
+            StructField("COBORROWER_NAME", StringType(), True),
+            StructField("COBORROWER_RELATION", StringType(), True),
+            StructField("COBORROWER_SEX", StringType(), True),
+            StructField("COBORROWER_SMOKER", StringType(), True),
+            StructField("COBORROWER_WEIGHT", StringType(), True),
+            StructField("COBORROWER_OCCUPATION", StringType(), True),
+            StructField("DATA_SOURCE", StringType(), True),
+            StructField("LEAD_ORIGIN_URL", StringType(), True),
+            StructField("MAILING_ID", StringType(), True),
+            StructField("SUSPECT_CAMPAIGN_ID", DecimalType(38, 0), True),
+            StructField("CONSUMER_DEBT", DoubleType(), True),
+            StructField("MORTGAGE_DEBT", StringType(), True),
+            StructField("UTM_CAMPAIGN", StringType(), True),
+            StructField("UTM_MEDIUM", StringType(), True),
+            StructField("UTM_SOURCE", StringType(), True),
+            StructField("REFERRING_URL", StringType(), True),
+            StructField("PCS_POLICIES_ID", DecimalType(38, 0), True),
+            StructField("CREATE_DATE", TimestampType(), True),
+            StructField("MODIFY_DATE", TimestampType(), True),
+            StructField("SOURCE_TABLE", StringType(), True),
+            StructField("IS_DELETED_SOURCE", StringType(), True),
+            StructField("EXP_DATE", TimestampType(), True),
+            StructField("SOURCE_TYPE", StringType(), True),
+            StructField("SOURCE_TYPE_ID", DecimalType(38, 0), True),
+            StructField("PRODUCT_TYPE", StringType(), True),
+            StructField("LEAD_ATTRIBUTES", StringType(), True),
+            StructField("CUSTODY_TARGET_AUDIENCE", StringType(), True),
+            StructField("SOURCE", StringType(), True),
+            StructField("PRODUCT_TYPE_ID", DecimalType(38, 0), True),
+            StructField("LEAD_SOURCE_ID", StringType(), True),
+            StructField("ORIGIN_SYSTEM_ID", StringType(), True),
+            StructField("ORIGIN_SYSTEM", StringType(), True),
+            StructField("ORIGIN_SYSTEM_ORIG", StringType(), True),
+            StructField("LEAD_INGESTION_METHOD", StringType(), True),
+            StructField("ETL_CREATED_DATE", TimestampType(), False),
+            StructField("ETL_LAST_UPDATE_DATE", TimestampType(), False),
+            StructField("CREATED_BY", StringType(), False),
+            StructField("TO_PROCESS", BooleanType(), False),
+            StructField("EDW_EXTERNAL_SOURCE_SYSTEM", StringType(), False),
+        ]),
+        "lead_xref": StructType([
+            StructField("STG_LCR_LEAD_XREF_KEY", StringType(), True),
+            StructField("LEAD_XREF_GUID", StringType(), True),
+            StructField("LEGACY_LEAD_ID", StringType(), True),
+            StructField("LEAD_CODE", StringType(), True),
+            StructField("LEAD_LEVEL_ID", StringType(), True),
+            StructField("LEAD_LEVEL", StringType(), True),
+            StructField("DATA_SOURCE_ID", StringType(), True),
+            StructField("LEVEL_DATE", TimestampType(), True),
+            StructField("CREATE_DATE", TimestampType(), True),
+            StructField("MODIFY_DATE", TimestampType(), True),
+            StructField("AVAILABLE_FOR_PURCHASE_IND", StringType(), True),
+            StructField("IS_DELETED_SOURCE", StringType(), True),
+            StructField("LEAD_LEVEL_ALIAS", StringType(), True),
+            StructField("ETL_CREATED_DATE", TimestampType(), False),
+            StructField("ETL_LAST_UPDATE_DATE", TimestampType(), False),
+            StructField("CREATED_BY", StringType(), False),
+            StructField("TO_PROCESS", BooleanType(), False),
+            StructField("EDW_EXTERNAL_SOURCE_SYSTEM", StringType(), False),
+        ]),
+        "lead_assignment": StructType([
+            StructField("STG_LCR_LEAD_ASSIGNMENT_KEY", StringType(), True),
+            StructField("LEAD_ASSIGNMENT_GUID", StringType(), True),
+            StructField("LEAD_XREF_GUID", StringType(), True),
+            StructField("AGENT_CODE", StringType(), True),
+            StructField("PURCHASE_DATE", TimestampType(), True),
+            StructField("PURCHASE_PRICE", DoubleType(), True),
+            StructField("ASSIGN_DATE", TimestampType(), True),
+            StructField("INACTIVE_IND", StringType(), True),
+            StructField("STATUS", StringType(), True),
+            StructField("AGENT_EXTUID", StringType(), True),
+            StructField("ALLOCATE_IND", StringType(), True),
+            StructField("COMMENTS", StringType(), True),
+            StructField("SFG_DIRECT_AGENT_ID", StringType(), True),
+            StructField("BASE_SHOP_OWNER_AGENT_ID", StringType(), True),
+            StructField("TOTAL_UPLINE_AGENT_CODES", StringType(), True),
+            StructField("UNPAID_IND", StringType(), True),
+            StructField("APP_COUNT", StringType(), True),
+            StructField("APP_APV", StringType(), True),
+            StructField("ACTUAL_APP_COUNT", StringType(), True),
+            StructField("ACTUAL_APV", StringType(), True),
+            StructField("CREATE_DATE", TimestampType(), True),
+            StructField("MODIFY_DATE", TimestampType(), True),
+            StructField("SOURCE_TABLE", StringType(), True),
+            StructField("METADATA", StringType(), True),
+            StructField("STATUS_DATE", TimestampType(), True),
+            StructField("IS_DELETED_SOURCE", BooleanType(), True),
+            StructField("ORDER_NUMBER", StringType(), True),
+            StructField("LEAD_STATUS_ID", StringType(), True),
+            StructField("LEAD_STATUS", StringType(), True),
+            StructField("HQ_PURCHASE_AMOUNT", DoubleType(), True),
+            StructField("LEAD_ORDER_SYSTEM_ID", StringType(), True),
+            StructField("LEAD_ORDER_SYSTEM", StringType(), True),
+            StructField("ORDER_SYSTEM_ID", StringType(), True),
+            StructField("ORDER_SYSTEM", StringType(), True),
+            StructField("ORDER_SYSTEM_ORIG", StringType(), True),
+            StructField("EXCLUSIVITY_END_DATE", TimestampType(), True),
+            StructField("ETL_CREATED_DATE", TimestampType(), False),
+            StructField("ETL_LAST_UPDATE_DATE", TimestampType(), False),
+            StructField("CREATED_BY", StringType(), False),
+            StructField("TO_PROCESS", BooleanType(), False),
+            StructField("EDW_EXTERNAL_SOURCE_SYSTEM", StringType(), False),
+        ]),
+    }
+
+    COLUMN_MAPPINGS: Dict[str, Dict[str, str]] = {
+        # ... All mapping dicts, verbatim, for each table ...
+        "lead": {
+            "leadguid": "LEAD_GUID",
+            "legacyleadid": "LEGACY_LEAD_ID",
+            "individ": "INDIV_ID",
+            "hhid": "HH_ID",
+            "addrid": "ADDR_ID",
+            "leadcode": "LEAD_CODE",
+            "leadtypeid": "LEAD_TYPE_ID",
+            "leadtype": "LEAD_TYPE",
+            "leadsource": "LEAD_SOURCE",
+            "leadcreatedate": "LEAD_CREATE_DATE",
+            "firstname": "FIRST_NAME",
+            "middlename": "MIDDLE_NAME",
+            "lastname": "LAST_NAME",
+            "suffix": "SUFFIX",
+            "birthdate": "BIRTH_DATE",
+            "age": "AGE",
+            "sex": "SEX",
+            "street1": "STREET_1",
+            "street2": "STREET_2",
+            "city": "CITY",
+            "stateid": "STATE_ID",
+            "state": "STATE",
+            "zip": "ZIP",
+            "zip5": "ZIP5",
+            "county": "COUNTY",
+            "country": "COUNTRY",
+            "phone": "PHONE",
+            "homephone": "HOME_PHONE",
+            "cellphone": "CELL_PHONE",
+            "workphone": "WORK_PHONE",
+            "donotcall": "DO_NOT_CALL",
+            "callerid": "CALLER_ID",
+            "email": "EMAIL",
+            "dynamiclead": "DYNAMIC_LEAD",
+            "prospectid": "PROSPECT_ID",
+            "extpartnerid": "EXT_PARTNER_ID",
+            "channelid": "CHANNEL_ID",
+            "channel": "CHANNEL",
+            "optsourceid": "OPT_SOURCE_ID",
+            "sourceid": "SOURCE_ID",
+            "subsourceid": "SUB_SOURCE_ID",
+            "sourceofreferral": "SOURCE_OF_REFERRAL",
+            "division": "DIVISION",
+            "leadsubsource": "LEAD_SUB_SOURCE",
+            "leadsubsourceid": "LEAD_SUB_SOURCE_ID",
+            "lender": "LENDER",
+            "loanamount": "LOAN_AMOUNT",
+            "loandate": "LOAN_DATE",
+            "diabetes": "DIABETES",
+            "healthproblems": "HEALTH_PROBLEMS",
+            "heartproblems": "HEART_PROBLEMS",
+            "height": "HEIGHT",
+            "highbpchol": "HIGH_BP_CHOL",
+            "isinsured": "IS_INSURED",
+            "smoker": "SMOKER",
+            "occupation": "OCCUPATION",
+            "spouse": "SPOUSE",
+            "coborrowerage": "COBORROWER_AGE",
+            "coborrowerbirthdate": "COBORROWER_BIRTH_DATE",
+            "coborrowerheight": "COBORROWER_HEIGHT",
+            "coborroweronmortgage": "COBORROWER_ON_MORTGAGE",
+            "coborrowername": "COBORROWER_NAME",
+            "coborrowerrelation": "COBORROWER_RELATION",
+            "coborrowersex": "COBORROWER_SEX",
+            "coborrowersmoker": "COBORROWER_SMOKER",
+            "coborrowerweight": "COBORROWER_WEIGHT",
+            "coborroweroccupation": "COBORROWER_OCCUPATION",
+            "datasource": "DATA_SOURCE",
+            "leadoriginurl": "LEAD_ORIGIN_URL",
+            "mailingid": "MAILING_ID",
+            "suspectcampaignid": "SUSPECT_CAMPAIGN_ID",
+            "consumerdebt": "CONSUMER_DEBT",
+            "mortgagedebt": "MORTGAGE_DEBT",
+            "utmcampaign": "UTM_CAMPAIGN",
+            "utmmedium": "UTM_MEDIUM",
+            "utmsource": "UTM_SOURCE",
+            "referringurl": "REFERRING_URL",
+            "pcspoliciesid": "PCS_POLICIES_ID",
+            "createdate": "CREATE_DATE",
+            "modifydate": "MODIFY_DATE",
+            "sourcetable": "SOURCE_TABLE",
+            "isdeletedsource": "IS_DELETED_SOURCE",
+            "expdate": "EXP_DATE",
+            "sourcetype": "SOURCE_TYPE",
+            "sourcetypeid": "SOURCE_TYPE_ID",
+            "producttype": "PRODUCT_TYPE",
+            "leadattributes": "LEAD_ATTRIBUTES",
+            "custodytargetaudience": "CUSTODY_TARGET_AUDIENCE",
+            "source": "SOURCE",
+            "producttypeid": "PRODUCT_TYPE_ID",
+            "leadsourceid": "LEAD_SOURCE_ID",
+            "originsystemid": "ORIGIN_SYSTEM_ID",
+            "originsystem": "ORIGIN_SYSTEM",
+            "originsystem_orig": "ORIGIN_SYSTEM_ORIG",
+            "leadingestionmethod": "LEAD_INGESTION_METHOD",
+        },
+        "lead_xref": {
+            "leadxrefguid": "LEAD_XREF_GUID",
+            "legacyleadid": "LEGACY_LEAD_ID",
+            "leadcode": "LEAD_CODE",
+            "leadlevelid": "LEAD_LEVEL_ID",
+            "leadlevel": "LEAD_LEVEL",
+            "datasourceid": "DATA_SOURCE_ID",
+            "leveldate": "LEVEL_DATE",
+            "createdate": "CREATE_DATE",
+            "modifydate": "MODIFY_DATE",
+            "availableforpurchaseind": "AVAILABLE_FOR_PURCHASE_IND",
+            "isdeletedsource": "IS_DELETED_SOURCE",
+            "leadlevelalias": "LEAD_LEVEL_ALIAS",
+        },
+        "lead_assignment": {
+            "leadassignmentguid": "LEAD_ASSIGNMENT_GUID",
+            "leadxrefguid": "LEAD_XREF_GUID",
+            "agentcode": "AGENT_CODE",
+            "purchasedate": "PURCHASE_DATE",
+            "purchaseprice": "PURCHASE_PRICE",
+            "assigndate": "ASSIGN_DATE",
+            "inactiveind": "INACTIVE_IND",
+            "status": "STATUS",
+            "agentextuid": "AGENT_EXTUID",
+            "allocateind": "ALLOCATE_IND",
+            "comments": "COMMENTS",
+            "sfgdirectagentid": "SFG_DIRECT_AGENT_ID",
+            "baseshopowneragentid": "BASE_SHOP_OWNER_AGENT_ID",
+            "totaluplineagentcodes": "TOTAL_UPLINE_AGENT_CODES",
+            "unpaidind": "UNPAID_IND",
+            "appcount": "APP_COUNT",
+            "appapv": "APP_APV",
+            "actualappcount": "ACTUAL_APP_COUNT",
+            "actualapv": "ACTUAL_APV",
+            "createdate": "CREATE_DATE",
+            "modifydate": "MODIFY_DATE",
+            "sourcetable": "SOURCE_TABLE",
+            "metadata": "METADATA",
+            "statusdate": "STATUS_DATE",
+            "isdeletedsource": "IS_DELETED_SOURCE",
+            "ordernumber": "ORDER_NUMBER",
+            "leadstatusid": "LEAD_STATUS_ID",
+            "leadstatus": "LEAD_STATUS",
+            "hqpurchaseamount": "HQ_PURCHASE_AMOUNT",
+            "leadordersystemid": "LEAD_ORDER_SYSTEM_ID",
+            "leadordersystem": "LEAD_ORDER_SYSTEM",
+            "ordersystemid": "ORDER_SYSTEM_ID",
+            "ordersystem": "ORDER_SYSTEM",
+            "ordersystemorig": "ORDER_SYSTEM_ORIG",
+            "exclusivityenddate": "EXCLUSIVITY_END_DATE",
+        },
+    }
+
+    JSON_COLUMNS = {
+        "lead_assignment": ["METADATA"],
+        "lead": ["LEAD_ATTRIBUTES"],
+        "lead_xref": [],
+    }
+
+    BOOLEAN_STRING_COLUMNS = {
+        "IS_DELETED_SOURCE",
+    }
+
+    @classmethod
+    def get_schema(cls, table_name: str):
+        return cls.TABLE_SCHEMAS[table_name]
+
+    @classmethod
+    def get_column_mapping(cls, table_name: str):
+        return cls.COLUMN_MAPPINGS[table_name]
+
+    @classmethod
+    def get_json_columns(cls, table_name: str):
+        return cls.JSON_COLUMNS.get(table_name, [])
+
+    @classmethod
+    def is_boolean_string_column(cls, col_name: str):
+        return col_name in cls.BOOLEAN_STRING_COLUMNS
